Add the ability to hide the main image link.
Sometimes users do not want to link to the images in the carousel,
as it is just meant for visual flair, not to show off photographs in
an interactive gallery.

This patch adds the parameter `linkimg`, which defaults to `"1"`, to
enable or disable the linking of the images in the carousel. When set
to `"0"` --- or any other Wordpress boolean false-equivalent --- the
plugin does not render the link.